+++ app/src/main/java/io/github/sds100/keymapper/actions/PerformActionsUseCase.kt
@@ -333,7 +333,11 @@ class PerformActionsUseCaseImpl(
             }
 
             is ActionData.Text -> {
-                imeInputEventInjector.inputText(action.text)
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                    accessibilityService.inputText(action.text)
+                } else {
+                    imeInputEventInjector.inputText(action.text)
+                }
                 result = Success(Unit)
             }